temp1

+++ temperature_conv/src/lib.rs
@@ -4,7 +4,7 @@ Formula
 */
 
 pub fn fahrenheit_to_celsius(f: f64) -> f64 {
-    (f - (32 as f64))*((5/9) as f64)
+    (f - 32.0)*(5.0/9.0)
 }
 
 /*
